move jmri/util/swing/ToggleOrPressButtonModel.java to java/src/jmri/util/com/sun/ToggleOrPressButtonModel.java, associated changes, add package-info.java in place of README
