--- java/src/jmri/util/com/sun/README
@@ -1,4 +1,0 @@
-This directory contains sample and tutorial code from Sun.
-
-See the comments at the top of each file for source and Sun's 
-licensing terms
