+++ java/src/jmri/util/com/sun/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * This directory contains sample and tutorial code from Sun.
+ * <p>
+ * See the comments at the top of each file for source and Sun's
+ * licensing terms.
+ * <p>
+ * We separately package this code
+ * <ol>
+ * <li>to make clear where the code comes from, and
+ * <li>to highlight that because this has been copied from Java code, it might be
+ *     quite brittle against changes from one Java version to another.
+ * </ol>
+ */
+// include empty DefaultAnnotation to avoid excessive recompilation
+@edu.umd.cs.findbugs.annotations.DefaultAnnotation(value={})
+package jmri.util.com.sun;
